[FIX] connector_elasticsearch: Uses requests as trasnport layer by default By default, the elastic search client uses the urllib3 transport which checks the certificates with the bundle provided by certifi.
However, in some cases, the certificate used is not available in certifi.

The elasticsearch client can also use requests as a transport layer. Requests is also based on certifi, but has the advantage that it can also be configured with the environment variable.

export REQUESTS_CA_CA_BUNDLE=/etc/ssl/certs/ca-bundle.crt

This change allows you to manage certificates at the system level, which is cleaner and simpler than adding configuration options to the code.