Fix recompute JSON data using SUDO
Use `sudo` because the recomputation can be triggered from everywhere
(eg: an update of a product in the stock) and is not granted
that the user triggering it has access to all required records
(eg: se.backend or related records needed to compute index values).
All in all, this is safe because the index data should always
be the same no matter the access rights of the user triggering this.