[FIX] connector_elasticsearch: ignore missing documents in delete
As ce6c775d375a states, delete is done in the same transaction as the export.
This means that an exception during delete fails the whole job.
If a document does not exist in the index, as far as delete is concerned,
the job is done. So we ignore these exceptions.
Also: there are plenty of trivial reasons for which this might happen.
The error is nonetheless logged.