connector_search_engine: force sequentiality for batch export per index
Force all the batchs for a given index to be exported one after the
other, inside  a (queue_job) chain.

Then if some reason a job fail to export, like a unreachable server
then it pause all the consecutives jobs.

It makes the errors easier to troubleshoot (less noise) and probably
help to not saturate the se server with lot of // update requests on
the same index.