[FIX] search_engine_image_thumbnail: Avoid error on view load
In the thumbnail size form a domain is computed to restrict the model's field for which the size is defined. The build process filters of type FSImage of relational field to image relation model. In this mechanism we check the comodel_name info to validate the related model if a comodel_name is defined on a field. In some cases (related), the comodel_name doesn't refer to an odoo model. (A bug?). We now ignore fields with a comodel_name not part of the registry