[FIX] connector_search_engine: export and delete on multi bindings
Fix error when calling the methods *export_record* or *delete_record* from
the *se.binding* model when called on a recordset with items from different
*se.backend*.

The *export* and *delete* methods involves the use of a *Backend Adapter* to
communicate with the target search engine. We then need to process the bindings
by backend to call the correct adapter and ensure at same time a batch process
of the requested operation for all the records linked to the same backend.